backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.10.0
Pillow==10.2.0
reportlab==4.0.8
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.3
httpx==0.26.0
lxml==5.1.0
PyMuPDF==1.23.8